Refacor project
