+++ lib/main.dart
@@ -1,22 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:training_app/ui/sign_in_page/sign_in_page.dart';
+
+import 'ui/home_page/home_page.dart';
 
 void main() {
   runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Dear Diary',
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const SignScreen(),
-    );
-  }
 }
