+++ lib/ui/home_page/home_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:training_app/ui/sign_in_page/sign_in_page.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Dear Diary',
+      debugShowCheckedModeBanner: false,
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const SignScreen(),
+    );
+  }
+}
